requirements.txt
fastapi
uvicorn
google-generativeai
numpy
pydantic
Pillow
aiogram
googlesearch-python
reportlab
duckduckgo-search
requests
aiohttp
markdown
edge-tts
